+++ lib/LibCrafts-1.0/Locales/frFR-Professions.lua
@@ -1,0 +1,16 @@
+local lib = LibCraftsGetLibrary()
+local name, locale, version = "Locales-Professions", "frFR", 1
+local module = --[[---@type LcLocaleModule]] lib:RegisterLocaleModule(name, locale, version)
+if not module then return end
+
+local L = module:GetStrings()
+L["Alchemy"] = "Alchimie"
+L["Blacksmithing"] = "Forge"
+L["Cooking"] = "Cuisine"
+L["Enchanting"] = "Enchantement"
+L["Engineering"] = "Ingénierie"
+L["First Aid"] = "Secourisme"
+L["Leatherworking"] = "Travail du cuir"
+L["Mining"] = "Minage"
+L["Poisons"] = "Poisons"
+L["Tailoring"] = "Couture"